Alpha 1.0.11
Added Local mdf

+++ UniversalWebApi/Models/User.cs
@@ -20,6 +20,6 @@ namespace UniversalWebApi.Models
     {
         public string Name { get; set; }
         public string Email { get; set; }
-        public int Age { get; set; }
+        public short Age { get; set; }
     }
 }
